# Reduce size of structs in cycle detector