--- a/src/core/cycle_detector_test.go
+++ b/src/core/cycle_detector_test.go
@@ -13,19 +13,19 @@ func TestCycleDetector(t *testing.T) {
 	targetC := ParseBuildLabel("//src:c", "")
 
 	t.Run("Add dep first dep", func(t *testing.T) {
-		err := cd.addDep(dependencyLink{from: targetA, to: targetB})
+		err := cd.addDep(dependencyLink{from: &targetA, to: &targetB})
 		assert.NoError(t, err)
-		assert.ElementsMatch(t, cd.deps[targetA], []BuildLabel{targetB})
+		assert.ElementsMatch(t, cd.deps[&targetA], []*BuildLabel{&targetB})
 	})
 
 	t.Run("Add second dep", func(t *testing.T) {
-		err := cd.addDep(dependencyLink{from: targetA, to: targetC})
+		err := cd.addDep(dependencyLink{from: &targetA, to: &targetC})
 		assert.NoError(t, err)
-		assert.ElementsMatch(t, cd.deps[targetA], []BuildLabel{targetB, targetC})
+		assert.ElementsMatch(t, cd.deps[&targetA], []*BuildLabel{&targetB, &targetC})
 	})
 
 	t.Run("Add cycle", func(t *testing.T) {
-		err := cd.addDep(dependencyLink{from: targetC, to: targetA})
+		err := cd.addDep(dependencyLink{from: &targetC, to: &targetA})
 		assert.Error(t, err, "didn't detect cycle from :c -> :a")
 	})
 }