Uses pointers to the `BuildTarget.Label` instead of a copy. This has a twofold effect as we stored these labels in the channel as well as in the map.    


Also the benchmark looks about 3x faster:

Before:
```
cpu: Intel(R) Xeon(R) E-2136 CPU @ 3.30GHz
BenchmarkCycles
BenchmarkCycles/Small
BenchmarkCycles/Small-12         	   59798	     19901 ns/op	      1508 links/ms	    4320 B/op	      24 allocs/op
BenchmarkCycles/Medium
BenchmarkCycles/Medium-12        	    7424	    171838 ns/op	      4542 links/ms	  112320 B/op	     624 allocs/op
BenchmarkCycles/Large
BenchmarkCycles/Large-12         	     225	   4975127 ns/op	      3927 links/ms	 3850493 B/op	   15724 allocs/op
```

After:
```
BenchmarkCycles
BenchmarkCycles/Small
BenchmarkCycles/Small-12                  109609             11853 ns/op              2531 links/ms          720 B/op         24 allocs/op
BenchmarkCycles/Medium
BenchmarkCycles/Medium-12                  19951             59710 ns/op             13066 links/ms        18720 B/op        624 allocs/op
BenchmarkCycles/Large
BenchmarkCycles/Large-12                     687           1711116 ns/op             11419 links/ms       943490 B/op      15681 allocs/op
```